test: move SyncBase._gather into tests
